restrict pages

+++ insta/views.py
@@ -1,6 +1,8 @@
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from .models import Image, Comment, Tag
 from .forms import ImageForm
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
@@ -27,10 +29,15 @@ def loginPage(request):
         else:
             messages.error(request, 'Invalid username or Password')
 
-
     context = {}
     return render(request, 'insta/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+@login_required(login_url='login')
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     images = Image.objects.filter(
@@ -47,6 +54,7 @@ def home(request):
     context = {'images': images, 'tags': tags, 'images_count': images_count, }
     return render(request, 'insta/home.html', context)
 
+@login_required(login_url='login')
 def post(request, pk):
     image = Image.objects.get(id=pk)
     comments = image.comment_set.all()
@@ -55,6 +63,7 @@ def post(request, pk):
     context = {'image': image, 'comments':comments, 'tags':tags, }
     return render(request, 'insta/posts.html', context)
 
+@login_required(login_url='login')
 def createPost(request): 
     form = ImageForm()
 
@@ -67,10 +76,15 @@ def createPost(request):
     context = {'form': form}
     return render(request, 'insta/post_form.html', context)
 
+@login_required(login_url='login')
 def updatePost(request, pk):
     image = Image.objects.get(id=pk)
     form = ImageForm(instance=image)
 
+    
+    if request.user != image.owner:
+        return HttpResponse('This method is restricted')
+        
     if request.method == 'POST':
         form =ImageForm(request.POST, request.FILES, instance=image)
         if form.is_valid():
@@ -80,8 +94,13 @@ def updatePost(request, pk):
     context = {'form':form}
     return render(request, 'insta/post_form.html', context)
 
+@login_required(login_url='login')
 def deletePost(request, pk):
     image = Image.objects.get(id=pk)
+
+    if request.user != image.owner:
+        return HttpResponse('This method is restricted')
+
     if request.method == 'POST':
         image.delete()
         return redirect('home')
